Fix the state count and take care of rounding issues in printing

+++ Ukelele Cocoa/UKKeyboardPrintView.m
@@ -97,7 +97,7 @@
 }
 
 - (BOOL)knowsPageRange:(NSRangePointer)range {
-	range->length = ceil([[self subviews] count] / self.printingInfo.viewsPerPage);
+	range->length = ceil([[self subviews] count] / (CGFloat)self.printingInfo.viewsPerPage);
 	return YES;
 }
